Removing unnecessary option and fixing conversion
First of all, thanks for this excellent template.

I've applied two small fixes.

The first one is removing the unnecessary `-u Mbps` option in the command. As per the docs, this option is only applicable for the `human-readable` output.

The second one is a fix in the download and upload conversion. The output bandwidth for both download and upload is returned in bytes. In order to have it in megabit per second, which matches the suffix `Mb/s` used in the template, we need to divide it by `125000` instead of `100000`.